fix python IDE autocompletion and imports
